Changed from IP addresses to MAC addresses
Should be solid.  Currently working on the live updating, not sure how
it will go, as I'm still trying to work out how to do so with a bar
graph and JSON connection (either one seems easier with either another
source of data, or a different type of graph, it's just combining the
two that isn't going so well).